# Searcher: support precise lang filters