--- a/cmd/searcher/internal/search/search.go
+++ b/cmd/searcher/internal/search/search.go
@@ -30,7 +30,7 @@ import (
 )
 
 const (
-	// numWorkers is how many concurrent readerGreps run in the case of
+	// numWorkers is how many concurrent goroutines run in the case of
 	// regexSearch, and the number of parallel workers in the case of
 	// structuralSearch.
 	numWorkers = 8
@@ -65,7 +65,7 @@ func (s *Service) search(ctx context.Context, p *protocol.Request, sender matchS
 		p.Commit.Attr(),
 		attribute.String("url", p.URL),
 		attribute.String("query", p.Query.String()),
-		attribute.StringSlice("languages", p.Languages),
+		attribute.StringSlice("languages", p.IncludeLangs),
 		attribute.Bool("isCaseSensitive", p.IsCaseSensitive),
 		attribute.Bool("pathPatternsAreCaseSensitive", p.PathPatternsAreCaseSensitive),
 		attribute.Int("limit", p.Limit),
@@ -101,7 +101,7 @@ func (s *Service) search(ctx context.Context, p *protocol.Request, sender matchS
 			log.String("commit", string(p.Commit)),
 			log.String("query", p.String()),
 			log.Bool("isStructuralPat", p.IsStructuralPat),
-			log.Strings("languages", p.Languages),
+			log.Strings("languages", p.IncludeLangs),
 			log.Bool("isCaseSensitive", p.IsCaseSensitive),
 			log.Bool("patternMatchesContent", p.PatternMatchesContent),
 			log.Bool("patternMatchesPath", p.PatternMatchesPath),
@@ -210,7 +210,8 @@ func validateParams(r *protocol.Request) error {
 	}
 
 	if p, ok := r.Query.(*protocol.PatternNode); ok {
-		if p.Value == "" && r.ExcludePattern == "" && len(r.IncludePatterns) == 0 {
+		if p.Value == "" && r.ExcludePaths == "" && len(r.IncludePaths) == 0 &&
+			len(r.IncludeLangs) == 0 && len(r.ExcludeLangs) == 0 {
 			return errors.New("At least one of pattern and include/exclude patterns must be non-empty")
 		}
 	}