--- a/cmd/searcher/internal/search/search_regex.go
+++ b/cmd/searcher/internal/search/search_regex.go
@@ -88,21 +88,6 @@ func regexSearch(
 		files = zf.Files
 	)
 
-	if _, ok := m.(allMatchTree); ok || (patternMatchesPaths && !patternMatchesContent) {
-		// Fast path for only matching file paths (or with a nil pattern, which matches all files,
-		// so is effectively matching only on file paths).
-		for _, f := range files {
-			if pm.Matches(f.Name) && m.MatchesString(f.Name) {
-				if ctx.Err() != nil {
-					return ctx.Err()
-				}
-				fm := protocol.FileMatch{Path: f.Name}
-				sender.Send(fm)
-			}
-		}
-		return nil
-	}
-
 	var (
 		lastFileIdx   = atomic.NewInt32(-1)
 		filesSkipped  atomic.Uint32
@@ -122,11 +107,7 @@ func regexSearch(
 
 	// Start workers. They read from files and write to matches.
 	for i := 0; i < numWorkers; i++ {
-		// transformBuf is reused between file searches to avoid
-		// re-allocating. It is only used if we need to transform the input
-		// before matching. For example we lower case the input in the case of
-		// ignoreCase.
-		var transformBuf []byte
+		l := fileLoader{zf: zf, isCaseSensitive: isCaseSensitive}
 		g.Go(func() error {
 			for !contextCanceled.Load() {
 				idx := int(lastFileIdx.Inc())
@@ -143,33 +124,30 @@ func regexSearch(
 				}
 				filesSearched.Inc()
 
-				// fileMatchBuf is what we run match on, fileBuf is the original
-				// data (for Preview).
-				fileBuf := zf.DataFor(f)
-				fileMatchBuf := fileBuf
-				if !isCaseSensitive {
-					// If we are ignoring case, we transform the input instead of
-					// relying on the regular expression engine which can be
-					// slow. compilePattern has already lowercased the pattern. We also
-					// trade some correctness for perf by using a non-utf8 aware
-					// lowercase function.
-					if transformBuf == nil {
-						transformBuf = make([]byte, zf.MaxLen)
-					}
-					fileMatchBuf = transformBuf[:len(fileBuf)]
-					casetransform.BytesToLowerASCII(fileMatchBuf, fileBuf)
+				match := false
+				fm := protocol.FileMatch{
+					Path:     f.Name,
+					LimitHit: false,
 				}
 
-				// find limit+1 matches so we know whether we hit the limit
-				match, locs := m.MatchesFile(fileMatchBuf, sender.Remaining()+1)
-				fm := locsToFileMatch(fileBuf, f.Name, locs, contextLines)
-				if !match && patternMatchesPaths {
-					// Try matching against the file path.
+				if patternMatchesPaths {
 					match = m.MatchesString(f.Name)
-					if match {
-						fm.Path = f.Name
+				}
+
+				if !match && patternMatchesContent {
+					if _, ok := m.(allMatchTree); ok {
+						// Avoid loading the file is this pattern always matches
+						match = true
+					} else {
+						l.load(f)
+
+						// find limit+1 matches so we know whether we hit the limit
+						var locs [][]int
+						match, locs = m.MatchesFile(l.fileMatchBuf, sender.Remaining()+1)
+						fm = locsToFileMatch(l.fileBuf, f.Name, locs, contextLines)
 					}
 				}
+
 				if match {
 					sender.Send(fm)
 				}
@@ -191,6 +169,43 @@ func regexSearch(
 	)
 
 	return err
+}
+
+// fileLoader loads files from the zipfile. It keeps a reference to the last
+// file it loaded, in case it's requested twice (for example first from language
+// detection, then pattern matching).
+type fileLoader struct {
+	zf              *zipFile
+	isCaseSensitive bool
+
+	// fileMatchBuf is what we run match on, fileBuf is the original data
+	// (used for the content preview).
+	fileBuf      []byte
+	fileMatchBuf []byte
+
+	// scratchBuf is reused between file searches to avoid
+	// re-allocating. It is only used if we need to transform the input
+	// before matching. For example we lower case the input in the case of
+	// ignoreCase.
+	scratchBuf []byte
+}
+
+func (l *fileLoader) load(f *srcFile) {
+	l.fileBuf = l.zf.DataFor(f)
+	l.fileMatchBuf = l.fileBuf
+
+	if !l.isCaseSensitive {
+		// If we are ignoring case, we transform the input instead of
+		// relying on the regular expression engine which can be
+		// slow. compilePattern has already lowercased the pattern. We also
+		// trade some correctness for perf by using a non-utf8 aware
+		// lowercase function.
+		if l.scratchBuf == nil {
+			l.scratchBuf = make([]byte, l.zf.MaxLen)
+		}
+		l.fileMatchBuf = l.scratchBuf[:len(l.fileBuf)]
+		casetransform.BytesToLowerASCII(l.fileMatchBuf, l.fileBuf)
+	}
 }
 
 // readAll will read r until EOF into b. It returns the number of bytes