--- a/internal/search/job/jobutil/BUILD.bazel
+++ b/internal/search/job/jobutil/BUILD.bazel
@@ -59,6 +59,7 @@ go_library(
         "//lib/errors",
         "//lib/iterator",
         "//schema",
+        "@com_github_go_enry_go_enry_v2//:go-enry",
         "@com_github_grafana_regexp//:regexp",
         "@com_github_sourcegraph_conc//pool",
         "@com_github_sourcegraph_log//:log",