… cmd/searcher/internal/search/BUILD.bazel
… cmd/searcher/internal/search/hybrid_test.go
… cmd/searcher/internal/search/pathmatch_test.go
… cmd/searcher/internal/search/search_regex_test.go
… cmd/searcher/internal/search/search_test.go
… cmd/searcher/protocol/searcher_test.go
… dev/gqltest/search_test.go
… internal/database/repos_test.go
… internal/search/job/jobutil/filter_file_contains_test.go
… internal/search/job/jobutil/job.go
… internal/search/job/jobutil/job_test.go
… internal/search/repos/repos_test.go
… internal/search/searcher/client.go
… internal/searcher/v1/searcher.pb.go