This PR adds support for the `search-content-based-lang-detection` feature flag
to unindexed search. It introduces a new type of matcher called `langMatcher`
that uses go-enry to determine a file's language.

The PR is big, but it's broken into commits. Main pieces:
* Refactor `search_regex` to introduce a `fileLoader` object
* Add language matching using go-enry
* Introduce new `IncludeLangs` and `ExcludeLangs` fields so we can pipe
language filters directly to searcher

Most file changes are just due renaming `IncludePatterns` -> `IncludePaths`.

## Test plan

Added unit tests and benchmarks