@jtibshirani 2024-02-14
Here are the searcher benchmarks before vs. after the change. I grouped each benchmark type together. Significant differences noted with `*...*`.

<details>
  <summary>Benchmark comparison </summary>

```
BenchmarkSearchRegex_large_fixed-10                  	     219	   5372972 ns/op	15692829 B/op	     243 allocs/op
BenchmarkSearchRegex_large_fixed-10                  	     216	   5427758 ns/op	15700592 B/op	     244 allocs/op

BenchmarkSearchRegex_rare_fixed-10                   	     168	   7246439 ns/op	15693014 B/op	     657 allocs/op
BenchmarkSearchRegex_rare_fixed-10                   	     169	   7591177 ns/op	15637489 B/op	     661 allocs/op

BenchmarkSearchRegex_large_fixed_casesensitive-10    	     436	   2765823 ns/op	   14090 B/op	     194 allocs/op
BenchmarkSearchRegex_large_fixed_casesensitive-10    	     441	   2716651 ns/op	   14797 B/op	     195 allocs/op

BenchmarkSearchRegex_large_empty_pattern-10    	     231	   5122012 ns/op	16780675 B/op	     105 allocs/op
BenchmarkSearchRegex_large_empty_pattern-10          	     228	   5145705 ns/op	16786850 B/op	     107 allocs/op

*Before: N/A*
*BenchmarkSearchRegex_large_lang_filter_common-10    	     261	   4557803 ns/op	15655989 B/op	     320 allocs/op*

*Before: N/A*
*BenchmarkSearchRegex_large_lang_filter_uncommon-10    	     974	   1274682 ns/op	15612261 B/op	     348 allocs/op*

BenchmarkSearchRegex_large_re_dotstar-10             	       5	 259080342 ns/op	628672656 B/op	 6098823 allocs/op
BenchmarkSearchRegex_large_re_dotstar-10             	       5	 224910450 ns/op	628902502 B/op	 6098810 allocs/op

BenchmarkSearchRegex_large_re_common-10              	     199	   6241995 ns/op	 7503550 B/op	   68778 allocs/op
BenchmarkSearchRegex_large_re_common-10              	     222	   5202429 ns/op	 7500757 B/op	   68776 allocs/op

BenchmarkSearchRegex_large_re_anchor-10              	      19	  57122230 ns/op	 4921381 B/op	   50146 allocs/op
BenchmarkSearchRegex_large_re_anchor-10              	      20	  57729229 ns/op	 4934114 B/op	   50150 allocs/op

BenchmarkSearchRegex_large_capture_group-10          	       8	 137275490 ns/op	 3449010 B/op	   25261 allocs/op
BenchmarkSearchRegex_large_capture_group-10          	       7	 147519845 ns/op	 3465006 B/op	   25264 allocs/op

*BenchmarkSearchRegex_large_path/path_only-10         	    7681	    152459 ns/op	    5266 B/op	      72 allocs/op*
*BenchmarkSearchRegex_large_path/path_only-10         	    1627	    681527 ns/op	    8269 B/op	     113 allocs/op*

BenchmarkSearchRegex_large_path/content_only-10      	     171	   6916182 ns/op	   64785 B/op	     652 allocs/op
BenchmarkSearchRegex_large_path/content_only-10      	     172	   7277534 ns/op	   66118 B/op	     653 allocs/op

BenchmarkSearchRegex_large_path/both_path_and_content-10         	     169	   7093483 ns/op	   65127 B/op	     652 allocs/op
BenchmarkSearchRegex_large_path/both_path_and_content-10         	     172	   6946727 ns/op	   63743 B/op	     631 allocs/op

BenchmarkSearchRegex_small_fixed-10                              	    7720	    156705 ns/op	  309957 B/op	     128 allocs/op
BenchmarkSearchRegex_small_fixed-10                              	    8428	    142722 ns/op	  324529 B/op	     130 allocs/op

BenchmarkSearchRegex_small_fixed_casesensitive-10                	   12189	    100133 ns/op	    9670 B/op	     122 allocs/op
BenchmarkSearchRegex_small_fixed_casesensitive-10                	   12505	     95307 ns/op	   10277 B/op	     123 allocs/op

BenchmarkSearchRegex_small_re_dotstar-10                         	     436	   2730551 ns/op	 4108227 B/op	   40140 allocs/op
BenchmarkSearchRegex_small_re_dotstar-10                         	     517	   2273420 ns/op	 4126375 B/op	   40142 allocs/op

BenchmarkSearchRegex_small_re_common-10                          	   10000	    108978 ns/op	   48431 B/op	     530 allocs/op
BenchmarkSearchRegex_small_re_common-10                          	   10000	    101683 ns/op	   49068 B/op	     531 allocs/op

BenchmarkSearchRegex_small_re_anchor-10                          	    1408	    840718 ns/op	   25483 B/op	     334 allocs/op
BenchmarkSearchRegex_small_re_anchor-10                          	    1626	    726375 ns/op	   26301 B/op	     335 allocs/op

BenchmarkSearchRegex_small_capture_group-10                      	     684	   1816197 ns/op	   24139 B/op	     315 allocs/op
BenchmarkSearchRegex_small_capture_group-10                      	     759	   1574914 ns/op	   24780 B/op	     316 allocs/op
```

Notes:
* The language filtering overhead isn't significant in common case where languages have unambiguous extensions
* Path-only searches became slightly slower because I needed to remove a dedicated "fast path" optimization