@keegancsmith 2024-02-14
> Here are the searcher benchmarks before vs. after the change. I grouped each benchmark type together. Significant differences noted with `*...*`.

I encourage the use of the benchstat command, much easier to parse and find changes. You need to run your benchmarks with `-count` to collect a few samples so the statistical comparison is fair. But here are the results on my desktop machine:

```
name                                             old time/op    new time/op    delta
SearchRegex_large_fixed-32                         5.93ms ±10%    6.32ms ± 8%      ~     (p=0.310 n=5+5)
SearchRegex_rare_fixed-32                          5.65ms ± 9%    5.79ms ±10%      ~     (p=0.690 n=5+5)
SearchRegex_large_fixed_casesensitive-32           3.09ms ± 3%    3.19ms ± 1%      ~     (p=0.095 n=5+5)
SearchRegex_large_re_dotstar-32                     259ms ± 2%     258ms ± 3%      ~     (p=1.000 n=5+5)
SearchRegex_large_re_common-32                     7.27ms ± 8%    7.16ms ± 9%      ~     (p=0.548 n=5+5)
SearchRegex_large_re_anchor-32                     66.9ms ± 1%    67.3ms ± 2%      ~     (p=0.690 n=5+5)
SearchRegex_large_capture_group-32                  174ms ± 5%     181ms ±10%      ~     (p=0.310 n=5+5)
SearchRegex_large_path/path_only-32                 153µs ± 1%     390µs ± 3%  +155.04%  (p=0.008 n=5+5)
SearchRegex_large_path/content_only-32             12.7ms ±10%    11.5ms ± 8%    -9.53%  (p=0.032 n=5+5)
SearchRegex_large_path/both_path_and_content-32    11.4ms ±22%    11.2ms ± 8%      ~     (p=0.841 n=5+5)
SearchRegex_small_fixed-32                          147µs ± 1%     149µs ± 1%    +1.77%  (p=0.008 n=5+5)
SearchRegex_small_fixed_casesensitive-32           95.5µs ± 3%    92.4µs ±10%      ~     (p=0.310 n=5+5)
SearchRegex_small_re_dotstar-32                    4.34ms ± 1%    4.34ms ± 2%      ~     (p=0.841 n=5+5)
SearchRegex_small_re_common-32                      170µs ± 2%     171µs ± 1%      ~     (p=0.690 n=5+5)
SearchRegex_small_re_anchor-32                     1.13ms ±35%    1.49ms ± 5%      ~     (p=0.222 n=5+5)
SearchRegex_small_capture_group-32                 3.41ms ±10%    3.62ms ± 7%      ~     (p=0.151 n=5+5)

name                                             old alloc/op   new alloc/op   delta
SearchRegex_large_fixed-32                         15.7MB ± 0%    15.7MB ± 0%      ~     (p=0.841 n=5+5)
SearchRegex_rare_fixed-32                          15.7MB ± 0%    15.7MB ± 0%      ~     (p=0.151 n=5+5)
SearchRegex_large_fixed_casesensitive-32           14.4kB ± 4%    15.4kB ± 2%    +6.90%  (p=0.008 n=5+5)
SearchRegex_large_re_dotstar-32                     630MB ± 0%     630MB ± 0%      ~     (p=0.548 n=5+5)
SearchRegex_large_re_common-32                     7.70MB ± 0%    7.70MB ± 0%      ~     (p=0.841 n=5+5)
SearchRegex_large_re_anchor-32                     4.95MB ± 0%    4.95MB ± 1%      ~     (p=0.690 n=5+5)
SearchRegex_large_capture_group-32                 3.46MB ± 1%    3.45MB ± 1%      ~     (p=0.690 n=5+5)
SearchRegex_large_path/path_only-32                5.27kB ± 0%    9.02kB ± 1%   +71.02%  (p=0.008 n=5+5)
SearchRegex_large_path/content_only-32             67.3kB ± 4%    68.2kB ± 4%      ~     (p=0.841 n=5+5)
SearchRegex_large_path/both_path_and_content-32    66.7kB ± 3%    66.8kB ± 4%      ~     (p=1.000 n=5+5)
SearchRegex_small_fixed-32                          473kB ± 0%     474kB ± 0%    +0.23%  (p=0.008 n=5+5)
SearchRegex_small_fixed_casesensitive-32           9.90kB ± 0%   11.04kB ± 0%   +11.52%  (p=0.008 n=5+5)
SearchRegex_small_re_dotstar-32                    4.17MB ± 0%    4.17MB ± 0%      ~     (p=0.095 n=5+5)
SearchRegex_small_re_common-32                     50.2kB ± 0%    51.4kB ± 0%    +2.40%  (p=0.008 n=5+5)
SearchRegex_small_re_anchor-32                     25.8kB ± 1%    27.0kB ± 2%    +4.71%  (p=0.008 n=5+5)
SearchRegex_small_capture_group-32                 24.5kB ± 2%    25.8kB ± 2%    +5.20%  (p=0.008 n=5+5)

name                                             old allocs/op  new allocs/op  delta
SearchRegex_large_fixed-32                            237 ± 1%       254 ± 1%    +7.25%  (p=0.008 n=5+5)
SearchRegex_rare_fixed-32                             674 ± 0%       692 ± 0%    +2.67%  (p=0.016 n=5+4)
SearchRegex_large_fixed_casesensitive-32              194 ± 0%       211 ± 0%    +8.76%  (p=0.008 n=5+5)
SearchRegex_large_re_dotstar-32                     6.10M ± 0%     6.10M ± 0%      ~     (p=0.056 n=5+5)
SearchRegex_large_re_common-32                      68.9k ± 0%     68.9k ± 0%    +0.02%  (p=0.008 n=5+5)
SearchRegex_large_re_anchor-32                      50.2k ± 0%     50.2k ± 0%    +0.03%  (p=0.008 n=5+5)
SearchRegex_large_capture_group-32                  25.3k ± 0%     25.3k ± 0%      ~     (p=0.095 n=5+5)
SearchRegex_large_path/path_only-32                  72.0 ± 0%     129.0 ± 0%   +79.17%  (p=0.008 n=5+5)
SearchRegex_large_path/content_only-32                653 ± 0%       669 ± 0%    +2.54%  (p=0.008 n=5+5)
SearchRegex_large_path/both_path_and_content-32       653 ± 0%       647 ± 0%    -0.80%  (p=0.008 n=5+5)
SearchRegex_small_fixed-32                            136 ± 0%       153 ± 0%   +12.50%  (p=0.016 n=4+5)
SearchRegex_small_fixed_casesensitive-32              122 ± 0%       139 ± 0%   +13.93%  (p=0.008 n=5+5)
SearchRegex_small_re_dotstar-32                     40.1k ± 0%     40.2k ± 0%    +0.04%  (p=0.016 n=4+5)
SearchRegex_small_re_common-32                        530 ± 0%       547 ± 0%    +3.21%  (p=0.008 n=5+5)
SearchRegex_small_re_anchor-32                        334 ± 0%       351 ± 0%    +5.09%  (p=0.008 n=5+5)
SearchRegex_small_capture_group-32                    315 ± 0%       332 ± 0%    +5.40%  (p=0.008 n=5+5)
```

It seems the main differences the benchmarks are noticing are all around allocations/memory usage. However, the significant changes here are all on something which in real terms is small amounts of memory. So seems acceptable.