@jtibshirani 2024-02-19
Thanks for the reviews! I tried to respond or fix everything. Happy to push any follow-ups if something's resolved.

Here are my final results from benchstat. They look good to me -- no significant regressions, and the new "lang filter" benchmarks have decent latency and allocs.

<details>

```
goos: darwin
goarch: arm64
pkg: github.com/sourcegraph/sourcegraph/cmd/searcher/internal/search
                                                │   old.txt    │               new.txt                │
                                                │    sec/op    │    sec/op     vs base                │
SearchRegex_large_fixed-10                        5.280m ± ∞ ¹   5.354m ± ∞ ¹         ~ (p=0.056 n=5)
SearchRegex_rare_fixed-10                         6.973m ± ∞ ¹   7.032m ± ∞ ¹    +0.85% (p=0.008 n=5)
SearchRegex_large_fixed_casesensitive-10          2.646m ± ∞ ¹   2.776m ± ∞ ¹    +4.94% (p=0.008 n=5)
SearchRegex_large_re_dotstar-10                   221.0m ± ∞ ¹   224.6m ± ∞ ¹         ~ (p=0.151 n=5)
SearchRegex_large_re_common-10                    5.203m ± ∞ ¹   5.345m ± ∞ ¹    +2.74% (p=0.008 n=5)
SearchRegex_large_re_anchor-10                    56.48m ± ∞ ¹   56.97m ± ∞ ¹    +0.86% (p=0.016 n=5)
SearchRegex_large_capture_group-10                139.5m ± ∞ ¹   155.4m ± ∞ ¹   +11.39% (p=0.032 n=5)
SearchRegex_large_path/path_only-10               153.0µ ± ∞ ¹   752.9µ ± ∞ ¹  +392.00% (p=0.008 n=5)
SearchRegex_large_path/content_only-10            7.015m ± ∞ ¹   6.934m ± ∞ ¹    -1.16% (p=0.032 n=5)
SearchRegex_large_path/both_path_and_content-10   6.917m ± ∞ ¹   6.960m ± ∞ ¹         ~ (p=0.222 n=5)
SearchRegex_small_fixed-10                        141.1µ ± ∞ ¹   143.1µ ± ∞ ¹    +1.41% (p=0.008 n=5)
SearchRegex_small_fixed_casesensitive-10          98.11µ ± ∞ ¹   98.83µ ± ∞ ¹    +0.74% (p=0.008 n=5)
SearchRegex_small_re_dotstar-10                   2.267m ± ∞ ¹   2.256m ± ∞ ¹         ~ (p=0.222 n=5)
SearchRegex_small_re_common-10                    102.8µ ± ∞ ¹   102.8µ ± ∞ ¹         ~ (p=0.841 n=5)
SearchRegex_small_re_anchor-10                    725.8µ ± ∞ ¹   726.0µ ± ∞ ¹         ~ (p=1.000 n=5)
SearchRegex_small_capture_group-10                1.584m ± ∞ ¹   1.575m ± ∞ ¹         ~ (p=0.310 n=5)
SearchRegex_large_empty_pattern-10                               5.199m ± ∞ ¹
SearchRegex_large_lang_filter_common-10                          4.375m ± ∞ ¹
SearchRegex_large_lang_filter_uncommon-10                        1.213m ± ∞ ¹
geomean                                           2.895m         3.209m         +12.11%
¹ need >= 6 samples for confidence interval at level 0.95

                                                │    old.txt    │                new.txt                │
                                                │     B/op      │      B/op       vs base               │
SearchRegex_large_fixed-10                        14.97Mi ± ∞ ¹    14.97Mi ± ∞ ¹        ~ (p=0.841 n=5)
SearchRegex_rare_fixed-10                         14.97Mi ± ∞ ¹    14.97Mi ± ∞ ¹        ~ (p=0.095 n=5)
SearchRegex_large_fixed_casesensitive-10          13.76Ki ± ∞ ¹    14.84Ki ± ∞ ¹   +7.79% (p=0.008 n=5)
SearchRegex_large_re_dotstar-10                   599.8Mi ± ∞ ¹    599.7Mi ± ∞ ¹        ~ (p=0.841 n=5)
SearchRegex_large_re_common-10                    7.145Mi ± ∞ ¹    7.148Mi ± ∞ ¹        ~ (p=0.222 n=5)
SearchRegex_large_re_anchor-10                    4.692Mi ± ∞ ¹    4.695Mi ± ∞ ¹        ~ (p=1.000 n=5)
SearchRegex_large_capture_group-10                3.298Mi ± ∞ ¹    3.294Mi ± ∞ ¹        ~ (p=1.000 n=5)
SearchRegex_large_path/path_only-10               5.146Ki ± ∞ ¹    8.592Ki ± ∞ ¹  +66.94% (p=0.008 n=5)
SearchRegex_large_path/content_only-10            63.27Ki ± ∞ ¹    66.71Ki ± ∞ ¹   +5.44% (p=0.032 n=5)
SearchRegex_large_path/both_path_and_content-10   63.64Ki ± ∞ ¹    62.76Ki ± ∞ ¹   -1.38% (p=0.008 n=5)
SearchRegex_small_fixed-10                        321.9Ki ± ∞ ¹    346.1Ki ± ∞ ¹   +7.52% (p=0.008 n=5)
SearchRegex_small_fixed_casesensitive-10          9.505Ki ± ∞ ¹   10.592Ki ± ∞ ¹  +11.44% (p=0.008 n=5)
SearchRegex_small_re_dotstar-10                   3.934Mi ± ∞ ¹    3.934Mi ± ∞ ¹        ~ (p=0.421 n=5)
SearchRegex_small_re_common-10                    47.36Ki ± ∞ ¹    48.42Ki ± ∞ ¹   +2.24% (p=0.008 n=5)
SearchRegex_small_re_anchor-10                    25.01Ki ± ∞ ¹    26.15Ki ± ∞ ¹   +4.54% (p=0.008 n=5)
SearchRegex_small_capture_group-10                23.63Ki ± ∞ ¹    24.75Ki ± ∞ ¹   +4.75% (p=0.008 n=5)
SearchRegex_large_empty_pattern-10                                 16.01Mi ± ∞ ¹
SearchRegex_large_lang_filter_common-10                            14.93Mi ± ∞ ¹
SearchRegex_large_lang_filter_uncommon-10                          14.95Mi ± ∞ ¹
geomean                                           435.7Ki          805.0Ki         +5.92%
¹ need >= 6 samples for confidence interval at level 0.95

                                                │   old.txt    │               new.txt               │
                                                │  allocs/op   │  allocs/op    vs base               │
SearchRegex_large_fixed-10                         243.0 ± ∞ ¹    260.0 ± ∞ ¹   +7.00% (p=0.008 n=5)
SearchRegex_rare_fixed-10                          659.0 ± ∞ ¹    676.0 ± ∞ ¹   +2.58% (p=0.008 n=5)
SearchRegex_large_fixed_casesensitive-10           194.0 ± ∞ ¹    211.0 ± ∞ ¹   +8.76% (p=0.008 n=5)
SearchRegex_large_re_dotstar-10                   6.099M ± ∞ ¹   6.099M ± ∞ ¹        ~ (p=1.000 n=5)
SearchRegex_large_re_common-10                    68.77k ± ∞ ¹   68.79k ± ∞ ¹   +0.03% (p=0.008 n=5)
SearchRegex_large_re_anchor-10                    50.15k ± ∞ ¹   50.16k ± ∞ ¹   +0.03% (p=0.008 n=5)
SearchRegex_large_capture_group-10                25.27k ± ∞ ¹   25.28k ± ∞ ¹   +0.02% (p=0.008 n=5)
SearchRegex_large_path/path_only-10                72.00 ± ∞ ¹   129.00 ± ∞ ¹  +79.17% (p=0.008 n=5)
SearchRegex_large_path/content_only-10             652.0 ± ∞ ¹    669.0 ± ∞ ¹   +2.61% (p=0.008 n=5)
SearchRegex_large_path/both_path_and_content-10    652.0 ± ∞ ¹    647.0 ± ∞ ¹   -0.77% (p=0.008 n=5)
SearchRegex_small_fixed-10                         129.0 ± ∞ ¹    146.0 ± ∞ ¹  +13.18% (p=0.008 n=5)
SearchRegex_small_fixed_casesensitive-10           122.0 ± ∞ ¹    139.0 ± ∞ ¹  +13.93% (p=0.008 n=5)
SearchRegex_small_re_dotstar-10                   40.14k ± ∞ ¹   40.16k ± ∞ ¹   +0.04% (p=0.008 n=5)
SearchRegex_small_re_common-10                     530.0 ± ∞ ¹    547.0 ± ∞ ¹   +3.21% (p=0.008 n=5)
SearchRegex_small_re_anchor-10                     334.0 ± ∞ ¹    351.0 ± ∞ ¹   +5.09% (p=0.008 n=5)
SearchRegex_small_capture_group-10                 315.0 ± ∞ ¹    332.0 ± ∞ ¹   +5.40% (p=0.008 n=5)
SearchRegex_large_empty_pattern-10                                122.0 ± ∞ ¹
SearchRegex_large_lang_filter_common-10                           318.0 ± ∞ ¹
SearchRegex_large_lang_filter_uncommon-10                         348.0 ± ∞ ¹
geomean                                           1.844k         1.419k         +7.57%
¹ need >= 6 samples for confidence interval at level 0.95
```
</details>